move mempool mock into top-level mock package
